// apps/web-ele/apis/types/notice.d.ts
export type NoticePageResponseDto = {
  /** 任意合法数值 */
  [property: string]: any;
  /* 通知所关联的客户端页面信息 */
  clientPage: ClientPageDto;
  /* 创建时间 */
  createdAt: string;
  /* 启用的平台 */
  enablePlatform: number;
  /* 通知ID */
  id: number;
  /* 是否置顶 */
  isPinned?: boolean;
  /* 是否发布 */
  isPublished: boolean;
  /* 通知类型 */
  noticeType: 0 | 1 | 2 | 3;
  /* 排序权重（数值越大越靠前） */
  order?: number;
  /* 关联页面代码 */
  pageCode?: string;
  /* 通知弹窗背景图片URL */
  popupBackgroundImage?: string;
  /* 优先级 */
  priorityLevel: 0 | 1 | 2 | 3;
  /* 发布结束时间 */
  publishEndTime?: string;
  /* 发布开始时间 */
  publishStartTime?: string;
  /* 阅读次数 */
  readCount?: number;
  /* 是否弹窗显示 */
  showAsPopup?: boolean;
  /* 通知标题 */
  title: string;

  /* 更新时间 */
  updatedAt: string;
};

/**
 *  类型定义 [ClientPageDto]
 *  @来源 components.schemas
 *  @更新时间 2025-09-29 10:18:47
 */
export type ClientPageDto = {
  /** 任意合法数值 */
  [property: string]: any;
  /* 主键id */
  id: number;
  /* 页面编码（唯一标识） */
  pageCode: string;
  /* 页面名称 */
  pageName: string;

  /* 页面路径（URL路径） */
  pagePath: string;
};